+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## Fixed
+## [1.6.1] - 2025-07-27
+
+### Fixed
 
 - PR [#131](https://github.com/marinasundstrom/CheckedExceptions/pull/131) Redo: Prevent catch clause fix in try-wrapped lambda